Added Input Validation

+++ env.sh
@@ -1,5 +1,7 @@
 #!/usr/bin/env bash
 
+set -e
+
 timestamp() {
   date +"%T"
 }
@@ -10,15 +12,21 @@ export LOGFILE="${DIR}/__run-$(timestamp).log"
 export KUBECONFIG="${DIR}/kubeconfig"
 export HELM_HOME="${DIR}/helm-home"
 
-echo 'Setting up helm'
+if [ -z "$1" ]; then
+  echo "You have to specify your username as first argument, e.g. martin"
 
-rm -rf "${HELM_HOME}" 2>/dev/null
-kubectl create ns $1-tiller 2>/dev/null >> "${LOGFILE}"
-helm init --client-only >> "${LOGFILE}"
-helm plugin install https://github.com/rimusz/helm-tiller >> "${LOGFILE}"
-helm tiller start $1-tiller >> "${LOGFILE}"
-helm tiller env >> "${LOGFILE}"
+else 
+  echo 'Setting up helm'
 
-echo 'THIS CLUSTER IS LIVE'
-echo 'Running workloads:'
-helm ls
+  rm -rf "${HELM_HOME}" 2>/dev/null
+  kubectl create ns $1-tiller 2>/dev/null >> "${LOGFILE}"
+  helm init --client-only >> "${LOGFILE}"
+  helm plugin install https://github.com/rimusz/helm-tiller >> "${LOGFILE}"
+  helm tiller start-ci $1-tiller >> "${LOGFILE}"
+  source <(helm tiller env) 
+
+  echo 'THIS CLUSTER IS LIVE'
+  echo 'Running workloads:'
+  helm ls
+fi
+
